% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstmohi.R
\name{ls.e}
\alias{ls.e}
\title{Get emissivity}
\usage{
ls.e(b5, b4)
}
\arguments{
\item{b5}{Raster object}

\item{b4}{Raster object}
}
\value{
emissivity
}
\description{
Get emissivity
}
\examples{
ls.e(b5,b4)
}
