% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstmohi.R
\name{ls.lstK}
\alias{ls.lstK}
\title{Get land surface temperature (LST) in Kelvin}
\usage{
ls.lstK(b10, b5, b4)
}
\arguments{
\item{b10}{band 10/ thermal band of Landsat 8 image}

\item{b5}{band 5/ Near Infra-Red (NIR) band of Landsat 8 image}

\item{b4}{band 4/ Near Infra-Red (NIR) band of Landsat 8 image}
}
\value{
LST in Kelvin
}
\description{
Get land surface temperature (LST) in Kelvin
}
\examples{
ls.lst(b5,b4,b10)
}
