% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstmohi.R
\name{ls.ndvi}
\alias{ls.ndvi}
\title{Get Normalized Difference Vegetation Index (NDVI)}
\usage{
ls.ndvi(b5, b4)
}
\arguments{
\item{b5}{band 5/ Near Infra-Red (NIR) band of Landsat 8 image}

\item{b4}{band 4/ Near Infra-Red (NIR) band of Landsat 8 image}
}
\value{
ndvi
}
\description{
Get Normalized Difference Vegetation Index (NDVI)
}
\examples{
ls.ndvi(b5,b4)
}
