% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstmohi.R
\name{ls.tb}
\alias{ls.tb}
\title{Get Top of atmosphere brightness temperature (TB) in degree celcius}
\usage{
ls.tb(b10)
}
\arguments{
\item{b10}{band 10/ thermal band of Landsat 8 image}
}
\value{
TB
}
\description{
Get Top of atmosphere brightness temperature (TB) in degree celcius
}
\examples{
ls.tb(b10)
}
