% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstmohi.R
\name{ls.e}
\alias{ls.e}
\title{Get emissivity}
\usage{
ls.e(b5, b4)
}
\arguments{
\item{b5}{band 5/ Near Infra-Red (NIR) band of Landsat 8 image}

\item{b4}{band 4/ Near Infra-Red (NIR) band of Landsat 8 image}
}
\value{
emissivity
}
\description{
Get emissivity
}
\examples{
ls.e(b5,b4)
}
